+++ samples/openapi3/server/petstore/go/go-petstore/go/logger.go
@@ -11,23 +11,10 @@
 package petstoreserver
 
 import (
-	"log"
 	"net/http"
-	"time"
+	"github.com/go-chi/chi/v5/middleware"
 )
 
-func Logger(inner http.Handler, name string) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		start := time.Now()
-
-		inner.ServeHTTP(w, r)
-
-		log.Printf(
-			"%s %s %s %s",
-			r.Method,
-			r.RequestURI,
-			name,
-			time.Since(start),
-		)
-	})
+func Logger(inner http.Handler) http.Handler {
+	return middleware.Logger(inner)
 }
