+++ samples/server/others/go-server/no-body-path-params/go/logger.go
@@ -11,8 +11,8 @@
 package petstoreserver
 
 import (
-	"log"
 	"net/http"
+	"log"
 	"time"
 )
